+++ CHANGELOG.md
@@ -3,6 +3,10 @@ AYADN
 
 AYADN is a learning exercise: while it's fully functional, it's still very alpha (remember I'm a beginner with Ruby).  
 
+## 2014-01-18 v0.5.8a
+- bug fix (empty name in checkins)
+- feature: #nowplaying (iTunes + OS X only)
+
 ## 2014-01-09 v0.5.7a
 - quote a post (different from repost)